% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consonance.hutchinson1978.R
\name{consonance.hutchinson1978}
\alias{consonance.hutchinson1978}
\title{Consonance:}
\usage{
consonance.hutchinson1978(chord)
}
\arguments{
\item{chord}{A chord expressed as a vector of integers}
}
\value{
A tibble
}
\description{
Implements Hutchinson's 1978 roughness metric of consonance
}
