% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consonance.mulloy2022.R
\name{consonance.primes}
\alias{consonance.primes}
\title{Consonance: Prime Factors Sum}
\usage{
consonance.primes(chord)
}
\arguments{
\item{chord}{A chord expressed as an interval integers}
}
\value{
A tibble
}
\description{
Provides a simple metric of consonance based on the sum of prime factors
}
