% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consonance.stolzenburg2015.R
\name{consonance.stolzenburg2015}
\alias{consonance.stolzenburg2015}
\title{Consonance:}
\usage{
consonance.stolzenburg2015(chord)
}
\arguments{
\item{chord}{A chord expressed as an interval integers}

\item{observation_pitch}{The dominant observation point expressed as an integer}
}
\value{
A tibble
}
\description{
Implements Stolzenburg's 2015 periodicity metric of consonance
}
