% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmony.R
\name{harmony}
\alias{harmony}
\alias{h}
\title{Harmony}
\usage{
harmony(x, direction = +1, reference_tone = NULL, name = NULL)

h(x, direction = +1, reference_tone = NULL, name = NULL)
}
\arguments{
\item{x}{A note or chord expressed as an interval integer or vector of interval integers}

\item{direction}{Harmonic direction +1 is up and -1 is down}

\item{reference_tone}{The reference tone of the chord or larger context}

\item{name}{A custom name for the note or chord}
}
\value{
A tibble with interval, intervallic_name, name, affinity, brightness and magnitude
}
\description{
Provides the musical harmony metrics of a note or chord.
}
\details{
harmony(c(0,4,7)) # C Major
harmony(c(0,4,7),direction=1,name="C Major")
harmony(c(0+12,4,7),direction=-1,name="C Major 1st Inversion")
harmony(c(0+12,4+12,7),direction=1,name="C Major 2nd Inversion")
}
