% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.musical.field.theory-package.R
\docType{package}
\name{harmony.research-package}
\alias{harmony.research}
\alias{harmony.research-package}
\title{harmony.research: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
