% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melody.R
\name{melody}
\alias{melody}
\alias{m}
\title{Melody}
\usage{
melody(progression)

m(progression)
}
\arguments{
\item{progression}{A progression of chords as a multiple row tibble.}
}
\value{
A tibble
}
\description{
Provides the melodic metrics of a progression
}
