% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitch.R
\name{pitch}
\alias{pitch}
\alias{p}
\title{Pitch}
\usage{
pitch(x)

p(x)
}
\arguments{
\item{x}{A pitch expressed as an integer}
}
\value{
A tibble
}
\description{
Provides the fundamental attributes of a pitch in tonic-octave space
}
