% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consonance.stolzenburg2015.R
\name{consonance.stolzenburg2015}
\alias{consonance.stolzenburg2015}
\title{Consonance:}
\usage{
consonance.stolzenburg2015(chord)
}
\arguments{
\item{chord}{A chord expressed as a vector of integers}
}
\value{
A tibble
}
\description{
Implements a tonic-octave variation of Stolzenburg's 2015 tonic-only
consonance periodicity metric
}
