% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melody.R
\name{melody}
\alias{melody}
\alias{m}
\title{Melody}
\usage{
melody(progression, reference = NULL)

m(progression, reference = NULL)
}
\arguments{
\item{progression}{A progression of chords as a multiple row tibble.}

\item{reference}{The reference harmony for calculating potential energy}
}
\value{
A tibble
}
\description{
Provides the melodic metrics of a progression
}
