% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitch.R
\name{pitch}
\alias{pitch}
\alias{p}
\title{Pitch}
\usage{
pitch(x, root = 0)

p(x, root = 0)
}
\arguments{
\item{x}{A pitch expressed as an interval integer}

\item{root}{The aural root to evaluate the pitch}
}
\value{
A tibble
}
\description{
Provides the metrics of a pitch
}
